Changed the `SidebarsAPI` to `GroupsAPI` that includes exposures as well.
We want dbt exposures' tags and owners to show in the lineage API filters.
Currently, the filters in the lineage are read from the `sidebars` object in the report's data.
However, we don't want to add the exposures to the sidebars as it doesn't make sense.
Therefore, we changed the `SidebarsAPI` to `GroupsAPI` in which each item in a group (tags, owners, etc.) also references its resource type.
That way the report can exclude exposures from the sidebars but still include them in the lineage filter.