Allow specifying region_name for Boto3
When initialising a [Boto3 `Session`](https://boto3.amazonaws.com/v1/documentation/api/latest/guide/session.html#session-configurations) it's useful to be able to specify the AWS Region to which the client should connect.

Add an `--aws-region-name` option to the CLI which gets passed through to the `boto3` client used by the S3 client.

Context:

We're trying to use elementary to write to a Cloudflare R2 bucket, which implements the S3 API. However, the S3 endpoints for these buckets do not encode the region, and hence the `--s3-endpoint-url` flag isn't sufficient to be able to configure elementary to successfully access the buckets.

Without this flag, it means we need write an AWS Profile configuration to disk and pass that in to elementary in order to override the region, which isn't desirable in contexts where we integrate elementary (e.g. invoking it from Airflow).