disable elementary version check
checking the version is nice and all, but when running in production it just creates firewall log spam - we don't let traffic out anywhere, everything is explicitly allowed in kubernetes overlay network policy as well as an overarching firewall policy. Everytime elementary runs we see the spam trying to contact pypi to check for a newer version.

This PR creates a simple config to allow disabling the check.